+++ test/honeybadger/client_test.exs
@@ -1,0 +1,5 @@
+defmodule Honeybadger.ClientTest do
+  use ExUnit.Case, async: true
+
+  doctest Honeybadger.Client
+end