+++ CHANGELOG.md
@@ -9,6 +9,7 @@ adheres to [Semantic Versioning](http://semver.org/).
   (honeybadger-io/honeybadger-elixir#20).
 - Explicitly allow sending notices with strings and maps. For example, it is now
   possible to send a `RuntimeError` by calling `Honeybadger.notify("oops!")`.
+- Added Honeybadger test mix task which can be invoked using `mix honeybadger.test`
 
 ### Changed
 - Use the latest exception's stacktrace whenever `notify` is called from a